test: isolate preload query counting
